[TC-62]-refactor: 패키지 구조 개선
